refactor(multiplexer): extract shared infrastructure from tmux/zellij/herdr adapters
Extract quoteShellArg, buildOpencodeAttachCommand, and findBinary into
src/multiplexer/shared.ts. All three adapters now import from shared
instead of maintaining inline copies.

- shared.ts: 65 lines of shared infrastructure
- tmux: removed inline quoteShellArg, buildOpencodeAttachCommand, findBinary (-60 lines)
- zellij: removed inline quoteShellArg, buildOpencodeAttachCommand, findBinary (-36 lines)
- herdr: removed inline buildOpencodeAttachCommand, findBinary (-51 lines)
- findBinary simplified: removed -V verification step, derives logPrefix internally

Closes #676
